fix(eks): revisited cloudwatch alarms for cpu/disk abnormal operation
