+++ eks/cdk/app.py
@@ -1,17 +1,17 @@
 from aws_cdk import (
     core,
 )
-from eks import EKS
-from fis import CloudWatchAlarm
-from fis import FIS
 from ssm import DiskStressDocument
+from cw import CloudWatchAlarm
+from eks import EKS
+from fis import FIS
 
 props = {'namespace': 'fisworkshop'}
 app = core.App()
 
+ssm = DiskStressDocument(app, f"{props['namespace']}-ssm", props)
 eks = EKS(app, f"{props['namespace']}-eks", props)
-cw = CloudWatchAlarm(app, f"{props['namespace']}-cw", eks.outputs)
-disk_stress = DiskStressDocument(app, f"{props['namespace']}-ssm-diskstress", cw.outputs)
-fis = FIS(app, f"{props['namespace']}-fis", cw.outputs)
+alarm = CloudWatchAlarm(app, f"{props['namespace']}-cw", eks.outputs)
+fis = FIS(app, f"{props['namespace']}-fis", alarm.outputs)
 
 app.synth()
